+++ lib/src/scaffold.dart
@@ -98,7 +98,6 @@ class _CustomScaffoldState extends State<CustomScaffold> {
   @override
   Widget build(BuildContext context) {
     // TODO: possibly add a sticky bottom sheet in here...
-    //
     return Semantics(
       label: widget.semanticLabel,
       child: Scaffold(